requirements.txt
autogen-agentchat
autogen-core
autogen-ext
asyncio
dotenv
openai
tiktoken
autogen-ext[docker]
streamlit
requests
openpyxl
pandas
matplotlib
seaborn
fastapi
uvicorn